Added the midicontrol class which uses libdrumstick to provide footpedal support.
The Footpedal support allows adjusting of the tempo and skipping back by 10 seconds, currently it only supports the Line6 FBV pedal with non-default setup.

Also added support for audiopanorama, this currently moves the sound entirely to the left channel, unlike panning it moves the entire output it does not just switch to the left output. This currently defaults to on and cannot be turned off and only outputs to the left channel.